[UPDATE] distance thief police to integer

+++ migrations/20210401114006-create-game.js
@@ -58,7 +58,7 @@ module.exports = {
       },
       distanceThiefPolice:
       {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.INTEGER,
         allowNull: false,
         defaultValue: 25
       }
